Quadrotor flying but needs tuning work

+++ Rover/cpp/Rover.cpp
@@ -279,7 +279,7 @@ void Rover::transmitDataUDP()
 
 	pCntl.dataInt32.resize(4);
 	mMutex_cntl.lock();
-	Collection<uint16> cmds = mAttitudeThrustController.getLastMotorCmds();
+	Collection<uint8> cmds = mAttitudeThrustController.getLastMotorCmds();
 	mMutex_cntl.unlock();
 	for(int i=0; i<4; i++)
 		pCntl.dataInt32[i] = cmds[i];
